refactor numba
