+++ janitor/functions/conditional_join.py
@@ -1,6 +1,5 @@
 from __future__ import annotations
 import operator
-from enum import Enum
 from typing import Union, Any, Optional, Hashable, Literal
 import numpy as np
 import pandas as pd
@@ -17,7 +16,14 @@ from pandas.core.dtypes.common import (
 from pandas.core.reshape.merge import _MergeOperation
 
 from janitor.utils import check, check_column
-from janitor.functions.utils import _convert_to_numpy_array
+from janitor.functions.utils import (
+    _convert_to_numpy_array,
+    _JoinOperator,
+    _generic_func_cond_join,
+    _keep_output,
+    less_than_join_types,
+    greater_than_join_types,
+)
 
 
 @pf.register_dataframe_method
@@ -179,19 +185,6 @@ def conditional_join(
     )
 
 
-class _JoinOperator(Enum):
-    """
-    List of operators used in conditional_join.
-    """
-
-    GREATER_THAN = ">"
-    LESS_THAN = "<"
-    GREATER_THAN_OR_EQUAL = ">="
-    LESS_THAN_OR_EQUAL = "<="
-    STRICTLY_EQUAL = "=="
-    NOT_EQUAL = "!="
-
-
 operator_map = {
     _JoinOperator.STRICTLY_EQUAL.value: operator.eq,
     _JoinOperator.LESS_THAN.value: operator.lt,
@@ -199,16 +192,6 @@ operator_map = {
     _JoinOperator.GREATER_THAN.value: operator.gt,
     _JoinOperator.GREATER_THAN_OR_EQUAL.value: operator.ge,
     _JoinOperator.NOT_EQUAL.value: operator.ne,
-}
-
-
-less_than_join_types = {
-    _JoinOperator.LESS_THAN.value,
-    _JoinOperator.LESS_THAN_OR_EQUAL.value,
-}
-greater_than_join_types = {
-    _JoinOperator.GREATER_THAN.value,
-    _JoinOperator.GREATER_THAN_OR_EQUAL.value,
 }
 
 
@@ -454,14 +437,25 @@ def _conditional_join_compute(
             )
     else:
         left_on, right_on, op = conditions[0]
-        result = _generic_func_cond_join(
-            df[left_on],
-            right[right_on],
-            op,
-            False,
-            keep,
-            use_numba,
-        )
+        if use_numba:
+            from janitor.functions._numba import _numba_single_join
+
+            result = _numba_single_join(
+                left=df[left_on],
+                right=right[right_on],
+                op=op,
+                keep=keep,
+            )
+        else:
+            result = _generic_func_cond_join(
+                left=df[left_on],
+                right=right[right_on],
+                op=op,
+                multiple_conditions=False,
+                keep=keep,
+            )
+
+    # return result
 
     if result is None:
         result = np.array([], dtype=np.intp), np.array([], dtype=np.intp)
@@ -476,347 +470,6 @@ def _conditional_join_compute(
         right_columns,
         indicator,
     )
-
-
-def _keep_output(keep: str, left: np.ndarray, right: np.ndarray):
-    """return indices for left and right index based on the value of `keep`."""
-    if keep == "all":
-        return left, right
-    grouped = pd.Series(right).groupby(left)
-    if keep == "first":
-        grouped = grouped.min()
-        return grouped.index, grouped.array
-    grouped = grouped.max()
-    return grouped.index, grouped.array
-
-
-def _less_than_indices(
-    left: pd.Series,
-    right: pd.Series,
-    strict: bool,
-    keep: str,
-) -> tuple:
-    """
-    Use binary search to get indices where left
-    is less than or equal to right.
-
-    If strict is True, then only indices
-    where `left` is less than
-    (but not equal to) `right` are returned.
-
-    A tuple of integer indexes
-    for left and right is returned.
-    """
-
-    # no point going through all the hassle
-    if left.min() > right.max():
-        return None
-
-    any_nulls = left.isna()
-    if any_nulls.all():
-        return None
-    if any_nulls.any():
-        left = left[~any_nulls]
-    any_nulls = right.isna()
-    if any_nulls.all():
-        return None
-    if any_nulls.any():
-        right = right[~any_nulls]
-    any_nulls = any_nulls.any()
-    right_is_sorted = right.is_monotonic_increasing
-    if not right_is_sorted:
-        right = right.sort_values(kind="stable")
-
-    left_index = left.index._values
-    left = left._values
-    right_index = right.index._values
-    right = right._values
-
-    search_indices = right.searchsorted(left, side="left")
-
-    # if any of the positions in `search_indices`
-    # is equal to the length of `right_keys`
-    # that means the respective position in `left`
-    # has no values from `right` that are less than
-    # or equal, and should therefore be discarded
-    len_right = right.size
-    rows_equal = search_indices == len_right
-
-    if rows_equal.any():
-        left = left[~rows_equal]
-        left_index = left_index[~rows_equal]
-        search_indices = search_indices[~rows_equal]
-
-    # the idea here is that if there are any equal values
-    # shift to the right to the immediate next position
-    # that is not equal
-    if strict:
-        rows_equal = right[search_indices]
-        rows_equal = left == rows_equal
-        # replace positions where rows are equal
-        # with positions from searchsorted('right')
-        # positions from searchsorted('right') will never
-        # be equal and will be the furthermost in terms of position
-        # example : right -> [2, 2, 2, 3], and we need
-        # positions where values are not equal for 2;
-        # the furthermost will be 3, and searchsorted('right')
-        # will return position 3.
-        if rows_equal.any():
-            replacements = right.searchsorted(left, side="right")
-            # now we can safely replace values
-            # with strictly less than positions
-            search_indices = np.where(rows_equal, replacements, search_indices)
-        # check again if any of the values
-        # have become equal to length of right
-        # and get rid of them
-        rows_equal = search_indices == len_right
-
-        if rows_equal.any():
-            left = left[~rows_equal]
-            left_index = left_index[~rows_equal]
-            search_indices = search_indices[~rows_equal]
-
-        if not search_indices.size:
-            return None
-    if right_is_sorted and (keep == "first"):
-        if any_nulls:
-            return left_index, right_index[search_indices]
-        return left_index, search_indices
-    right = [right_index[ind:len_right] for ind in search_indices]
-    if keep == "first":
-        right = [arr.min() for arr in right]
-        return left_index, right
-    if keep == "last":
-        right = [arr.max() for arr in right]
-        return left_index, right
-    right = np.concatenate(right)
-    left = np.repeat(left_index, len_right - search_indices)
-    return left, right
-
-
-def _greater_than_indices(
-    left: pd.Series,
-    right: pd.Series,
-    strict: bool,
-    multiple_conditions: bool,
-    keep: str,
-) -> tuple:
-    """
-    Use binary search to get indices where left
-    is greater than or equal to right.
-
-    If strict is True, then only indices
-    where `left` is greater than
-    (but not equal to) `right` are returned.
-
-    if multiple_conditions is False, a tuple of integer indexes
-    for left and right is returned;
-    else a tuple of the index for left, right, as well
-    as the positions of left in right is returned.
-    """
-
-    # quick break, avoiding the hassle
-    if left.max() < right.min():
-        return None
-
-    any_nulls = left.isna()
-    if any_nulls.all():
-        return None
-    if any_nulls.any():
-        left = left[~any_nulls]
-    any_nulls = right.isna()
-    if any_nulls.all():
-        return None
-    if any_nulls.any():
-        right = right[~any_nulls]
-    any_nulls = any_nulls.any()
-    right_is_sorted = right.is_monotonic_increasing
-    if not right_is_sorted:
-        right = right.sort_values(kind="stable")
-
-    left_index = left.index._values
-    left = left._values
-    right_index = right.index._values
-    right = right._values
-
-    search_indices = right.searchsorted(left, side="right")
-    # if any of the positions in `search_indices`
-    # is equal to 0 (less than 1), it implies that
-    # left[position] is not greater than any value
-    # in right
-    rows_equal = search_indices < 1
-    if rows_equal.any():
-        left = left[~rows_equal]
-        left_index = left_index[~rows_equal]
-        search_indices = search_indices[~rows_equal]
-
-    # the idea here is that if there are any equal values
-    # shift downwards to the immediate next position
-    # that is not equal
-    if strict:
-        rows_equal = right[search_indices - 1]
-        rows_equal = left == rows_equal
-        # replace positions where rows are equal with
-        # searchsorted('left');
-        # however there can be scenarios where positions
-        # from searchsorted('left') would still be equal;
-        # in that case, we shift down by 1
-        if rows_equal.any():
-            replacements = right.searchsorted(left, side="left")
-            # return replacements
-            # `left` might result in values equal to len right
-            replacements = np.where(
-                replacements == right.size, replacements - 1, replacements
-            )
-            # now we can safely replace values
-            # with strictly greater than positions
-            search_indices = np.where(rows_equal, replacements, search_indices)
-        # any value less than 1 should be discarded
-        # since the lowest value for binary search
-        # with side='right' should be 1
-        rows_equal = search_indices < 1
-        if rows_equal.any():
-            left = left[~rows_equal]
-            left_index = left_index[~rows_equal]
-            search_indices = search_indices[~rows_equal]
-
-        if not search_indices.size:
-            return None
-
-    if multiple_conditions:
-        return left_index, right_index, search_indices
-    if right_is_sorted and (keep == "last"):
-        if any_nulls:
-            return left_index, right_index[search_indices - 1]
-        return left_index, search_indices - 1
-    right = [right_index[:ind] for ind in search_indices]
-    if keep == "first":
-        right = [arr.min() for arr in right]
-        return left_index, right
-    if keep == "last":
-        right = [arr.max() for arr in right]
-        return left_index, right
-    right = np.concatenate(right)
-    left = np.repeat(left_index, search_indices)
-    return left, right
-
-
-def _not_equal_indices(left: pd.Series, right: pd.Series, keep: str) -> tuple:
-    """
-    Use binary search to get indices where
-    `left` is exactly  not equal to `right`.
-
-    It is a combination of strictly less than
-    and strictly greater than indices.
-
-    A tuple of integer indexes for left and right
-    is returned.
-    """
-
-    dummy = np.array([], dtype=int)
-
-    # deal with nulls
-    l1_nulls = dummy
-    r1_nulls = dummy
-    l2_nulls = dummy
-    r2_nulls = dummy
-    any_left_nulls = left.isna()
-    any_right_nulls = right.isna()
-    if any_left_nulls.any():
-        l1_nulls = left.index[any_left_nulls.array]
-        l1_nulls = l1_nulls.to_numpy(copy=False)
-        r1_nulls = right.index
-        # avoid NAN duplicates
-        if any_right_nulls.any():
-            r1_nulls = r1_nulls[~any_right_nulls.array]
-        r1_nulls = r1_nulls.to_numpy(copy=False)
-        nulls_count = l1_nulls.size
-        # blow up nulls to match length of right
-        l1_nulls = np.tile(l1_nulls, r1_nulls.size)
-        # ensure length of right matches left
-        if nulls_count > 1:
-            r1_nulls = np.repeat(r1_nulls, nulls_count)
-    if any_right_nulls.any():
-        r2_nulls = right.index[any_right_nulls.array]
-        r2_nulls = r2_nulls.to_numpy(copy=False)
-        l2_nulls = left.index
-        nulls_count = r2_nulls.size
-        # blow up nulls to match length of left
-        r2_nulls = np.tile(r2_nulls, l2_nulls.size)
-        # ensure length of left matches right
-        if nulls_count > 1:
-            l2_nulls = np.repeat(l2_nulls, nulls_count)
-
-    l1_nulls = np.concatenate([l1_nulls, l2_nulls])
-    r1_nulls = np.concatenate([r1_nulls, r2_nulls])
-
-    outcome = _less_than_indices(left, right, strict=True, keep=keep)
-
-    if outcome is None:
-        lt_left = dummy
-        lt_right = dummy
-    else:
-        lt_left, lt_right = outcome
-
-    outcome = _greater_than_indices(
-        left, right, strict=True, multiple_conditions=False, keep=keep
-    )
-
-    if outcome is None:
-        gt_left = dummy
-        gt_right = dummy
-    else:
-        gt_left, gt_right = outcome
-
-    left = np.concatenate([lt_left, gt_left, l1_nulls])
-    right = np.concatenate([lt_right, gt_right, r1_nulls])
-
-    if (not left.size) & (not right.size):
-        return None
-    return _keep_output(keep, left, right)
-
-
-def _generic_func_cond_join(
-    left: pd.Series,
-    right: pd.Series,
-    op: str,
-    multiple_conditions: bool,
-    keep: str,
-    use_numba: bool,
-) -> tuple:
-    """
-    Generic function to call any of the individual functions
-    (_less_than_indices, _greater_than_indices,
-    or _not_equal_indices).
-    """
-    strict = False
-
-    if op in {
-        _JoinOperator.GREATER_THAN.value,
-        _JoinOperator.LESS_THAN.value,
-        _JoinOperator.NOT_EQUAL.value,
-    }:
-        strict = True
-
-    if use_numba:
-        if op in less_than_join_types:
-            op_code = 1
-        elif op in greater_than_join_types:
-            op_code = 0
-        else:
-            op_code = -1
-        from janitor.functions._numba import _numba_single_join
-
-        return _numba_single_join(left, right, strict, keep, op_code)
-
-    if op in less_than_join_types:
-        return _less_than_indices(left, right, strict, keep)
-    if op in greater_than_join_types:
-        return _greater_than_indices(
-            left, right, strict, multiple_conditions, keep
-        )
-    if op == _JoinOperator.NOT_EQUAL.value:
-        return _not_equal_indices(left, right, keep)
 
 
 def _generate_indices(
@@ -877,7 +530,6 @@ def _multiple_conditional_join_ne(
         op,
         multiple_conditions=False,
         keep="all",
-        use_numba=use_numba,
     )
     if indices is None:
         return None
@@ -953,7 +605,7 @@ def _multiple_conditional_join_le_lt(
     Returns a tuple of (df_index, right_index)
     """
     if use_numba:
-        from janitor.functions._numba import _numba_pair_le_lt
+        from janitor.functions._numba import _numba_dual_join
 
         pairs = [
             condition
@@ -996,9 +648,7 @@ def _multiple_conditional_join_le_lt(
                 pairs.extend(patch)
                 first, middle, last = pairs
                 pairs = [(first, middle), (first, last)]
-                indices = [
-                    _numba_pair_le_lt(df, right, pair) for pair in pairs
-                ]
+                indices = [_numba_dual_join(df, right, pair) for pair in pairs]
                 indices = [arr for arr in indices if arr is not None]
                 if not indices:
                     indices = None
@@ -1016,10 +666,9 @@ def _multiple_conditional_join_le_lt(
                     op,
                     multiple_conditions=False,
                     keep="all",
-                    use_numba=True,
                 )
         else:
-            indices = _numba_pair_le_lt(df, right, pairs)
+            indices = _numba_dual_join(df, right, pairs)
     else:
         # there is an opportunity for optimization for range joins
         # which is usually `lower_value < value < upper_value`
@@ -1100,8 +749,9 @@ def _multiple_conditional_join_le_lt(
                 op,
                 multiple_conditions=False,
                 keep="all",
-                use_numba=False,
             )
+
+    # return indices
 
     if not indices:
         return None
@@ -1153,41 +803,58 @@ def _range_indices(
     if any_nulls.any():
         right_c = right_c[~any_nulls]
 
-    strict = False
-    if op == _JoinOperator.GREATER_THAN.value:
-        strict = True
-
-    outcome = _greater_than_indices(
-        left_c,
-        right_c,
-        strict,
+    outcome = _generic_func_cond_join(
+        left=left_c,
+        right=right_c,
+        op=op,
         multiple_conditions=True,
         keep="all",
     )
+
+    # strict = False
+    # if op == _JoinOperator.GREATER_THAN.value:
+    #     strict = True
+
+    # outcome = _greater_than_indices(
+    #     left_c,
+    #     right_c,
+    #     strict,
+    #     multiple_conditions=True,
+    #     keep="all",
+    # )
 
     if outcome is None:
         return None
 
     left_index, right_index, search_indices = outcome
     left_on, right_on, op = second
-    right_c = right.loc[right_index, right_on]
-    left_c = df.loc[left_index, left_on]
-    strict = False
-    if op == _JoinOperator.LESS_THAN.value:
-        strict = True
-    op = operator_map[op]
+
+    # right_c = right.loc[right_index, right_on]
+    # left_c = df.loc[left_index, left_on]
+    # strict = False
+    # if op == _JoinOperator.LESS_THAN.value:
+    #     strict = True
+
     # fastpath
-    if right_c.is_monotonic_increasing:
-        outcome = _less_than_indices(
-            left_c,
-            right_c,
-            strict,
+    fastpath = right_c.is_monotonic_increasing
+    if fastpath:
+        outcome = _generic_func_cond_join(
+            left=df.loc[left_index, left_on],
+            right=right.loc[right_index, right_on],
+            op=op,
+            multiple_conditions=False,
             keep="first",
         )
+        # outcome = _less_than_indices(
+        #     left_c,
+        #     right_c,
+        #     strict,
+        #     multiple_conditions=False,
+        #     keep="first",
+        # )
         if outcome is None:
             return None
         left_c, pos = outcome
-        print(left_c, right_index, pos, left_index)
         if left_c.size < left_index.size:
             keep_rows = np.isin(left_index, left_c, assume_unique=True)
             search_indices = search_indices[keep_rows]
@@ -1195,6 +862,7 @@ def _range_indices(
     else:
         left_c = left_c._values
         right_c = right_c._values
+        op = operator_map[op]
         left_c, right_c = _convert_to_numpy_array(left_c, right_c)
         pos = np.copy(search_indices)
         counter = np.arange(left_c.size)
@@ -1231,16 +899,21 @@ def _range_indices(
 
     repeater = search_indices - pos
     if (repeater == 1).all():
-        # no point searching if the width is all 1
+        # no point running a comparison op
+        # if the width is all 1
         # this also implies that the intervals
         # do not overlap on the right side
         return left_index, right_index[pos]
+
     right_index = [
         right_index[start:end] for start, end in zip(pos, search_indices)
     ]
 
     right_index = np.concatenate(right_index)
     left_index = np.repeat(left_index, repeater)
+
+    if fastpath:
+        return left_index, right_index
     # here we search for actual positions
     # where left_c is </<= right_c
     # safe to index the arrays, since we are picking the positions
